fix: use bigint for migration counters (#806)
## What was changed
Updated the control-plane migration history so counter and generation
columns that `squawk` flagged as 32-bit integers now use `bigint`, and
kept the Drizzle schema plus migration snapshots aligned with that
historical SQL.

-
`packages/db/migrations/control-plane/0005_add-integration-automation-conversation-foundation.sql`:
changed `processor_generation` and `attempt_count` from `integer` to
`bigint`.
-
`packages/db/migrations/control-plane/0006_add-integration-connection-resources.sql`:
changed `total_count` from `integer` to `bigint`.
-
`packages/db/src/control-plane/schema/automation-conversation-delivery-tasks.ts`:
updated the Drizzle column definitions to `bigint(..., { mode: "number"
})` so the schema matches the migration history.
-
`packages/db/src/control-plane/schema/integration-connection-resource-states.ts`:
updated `totalCount` to the matching Drizzle `bigint` definition.
- `packages/db/migrations/control-plane/meta/0005_snapshot.json`:
updated the snapshot types for the delivery task columns to `bigint`.
- `packages/db/migrations/control-plane/meta/0006_snapshot.json`:
updated the snapshot types for the delivery task and resource state
columns to `bigint`.

Original symptom this addresses:

```text
warning[prefer-bigint-over-int]: Using 32-bit integer fields can result in hitting the max `int` limit.
migrations/control-plane/0005_add-integration-automation-conversation-foundation.sql:53
migrations/control-plane/0005_add-integration-automation-conversation-foundation.sql:55
migrations/control-plane/0006_add-integration-connection-resources.sql:6
```

## How to review
Start with
`packages/db/migrations/control-plane/0005_add-integration-automation-conversation-foundation.sql`
and
`packages/db/migrations/control-plane/0006_add-integration-connection-resources.sql`
to verify the lint-driven SQL change directly. Then read
`packages/db/src/control-plane/schema/automation-conversation-delivery-tasks.ts`
and
`packages/db/src/control-plane/schema/integration-connection-resource-states.ts`
to confirm the Drizzle definitions mirror those column types. Finish
with `packages/db/migrations/control-plane/meta/0005_snapshot.json` and
`packages/db/migrations/control-plane/meta/0006_snapshot.json` to check
the stored snapshots stay in sync with the schema history.

To validate locally, run `pnpm lint:sql` for the focused symptom or
`pnpm run ci` for the same full check sequence used before opening this
PR.

## What the implication was
New databases created from the historical control-plane migration chain
now get 64-bit storage for these counter and generation columns instead
of 32-bit integers, which removes the repo lint failure and avoids the
tighter integer ceiling that `squawk` warns about. Keeping the SQL,
Drizzle schema, and snapshots aligned also prevents future schema
regeneration from trying to reintroduce the same mismatch.

## Checks and tests performed
- `pnpm format`: verified repository formatting remained valid after the
bigint edits; expected outcome was no formatting changes left behind;
implemented by running the repo formatter across the worktree.
- `pnpm lint`: verified static analysis, SQL linting, docs linting,
OpenAPI drift checks, and Rust linting all passed with the bigint
changes; expected outcome was zero warnings/errors, including a clean
`pnpm lint:sql`; implemented by running the root lint workflow.
- `pnpm typecheck`: verified the Drizzle schema updates still typecheck
across the workspace; expected outcome was a clean TypeScript graph and
script typecheck; implemented by running the root typecheck workflow.
- `pnpm test`: verified the workspace test suites still pass after the
schema/migration alignment change; expected outcome was a clean turbo
test run; implemented by running the root test workflow.
- `pnpm run ci`: verified the full local CI mirror passes on the rebased
branch, including format check, lint, typecheck, Storybook build, and
tests; expected outcome was a clean end-to-end local CI run; implemented
by running the root CI script after rebasing onto `origin/main`.

+++ packages/db/migrations/control-plane/0006_add-integration-connection-resources.sql
@@ -3,7 +3,7 @@ CREATE TABLE "control_plane"."integration_connection_resource_states" (
 	"family_id" text NOT NULL,
 	"kind" text NOT NULL,
 	"sync_state" text DEFAULT 'never-synced' NOT NULL,
-	"total_count" integer DEFAULT 0 NOT NULL,
+	"total_count" bigint DEFAULT 0 NOT NULL,
 	"last_synced_at" timestamp with time zone,
 	"last_sync_started_at" timestamp with time zone,
 	"last_sync_finished_at" timestamp with time zone,
